@@ -14,20 +14,20 @@ class PhotoContainer extends Component {
     this.fetch(this.props.data);
   }
 
-  componentWillReceiveProps(nextProps) {
-    let routeChanged = nextProps.data !== this.props.data;
-    if (routeChanged) {
-      this.fetch(nextProps.data);
+  componentDidUpdate(nextProps) {
+    if (nextProps.data !== this.props.data) {
+      this.fetch(this.props.data);
     }
   }
 
   fetch = (query) => {
-    axios.get(`https://api.flickr.com/services/rest/?method=flickr.photos.search&api_key=${apiKey}&text=${query}&per_page=16&format=json&nojsoncallback=1`)
+    axios.get(`https://api.unsplash.com/search/photos/?client_id=${apiKey}&page=1&query=${query}&per_page=12`)
       .then(response => {
+        console.log(response.data.results)
         this.setState({
-          photos: response.data.photos.photo,
+          photos: response.data.results,
           loading: false
-        });
+        })
       })
       .catch(error => {
         console.log('Error fetching and parsing data', error);
@@ -36,7 +36,6 @@ class PhotoContainer extends Component {
 
   render() {
     const { title } = this.props;
-    console.log(title)
     return (
       <div className="photo-container">
         <h2>{ this.state.photos.length ? title : ''}</h2>
