@@ -2,7 +2,7 @@ import React from 'react';
 
 const Photo = (props) => (
   <li>
-    <img src={`https://farm${props.farm}.staticflickr.com/${props.server}/${props.id}_${props.secret}.jpg`} alt="" />
+    <img src={props.url} alt={props.alt} />
   </li>
 );
 
